Added replanning A* and FS

+++ flpa_node.h
@@ -12,14 +12,14 @@ struct FLPANode : virtual public Node
     int rhs;
     int minDst;
 
-    FLPANode(int x = 0, int y = 0, Node *p = nullptr, int g_ = 0, int H_ = 0, int ConflictsCount = 0, int hc_ = 0) :
+    FLPANode(int x = 0, int y = 0, Node *p = nullptr, int g_ = 0, double H_ = 0, int ConflictsCount = 0, int hc_ = 0) :
         Node(x, y, p, g_, H_, ConflictsCount), hc(hc_), futureConflictsCount(0) {}
 
     FLPANode(const Node &other) : Node(other) {}
 
     void setF() {
         minDst = std::min(g, rhs);
-        F = (minDst == CN_INFINITY) ? CN_INFINITY : minDst + H;
+        F = (minDst == CN_INFINITY) ? CN_INFINITY : double(minDst) + H;
     }
 
     virtual int convolution(int width, int height, bool withTime = true) const {
@@ -28,8 +28,8 @@ struct FLPANode : virtual public Node
     }
 
     bool operator< (const FLPANode &other) const {
-        return std::tuple<int, int, int, int, int>(F, minDst, -time, i, j) <
-                std::tuple<int, int, int, int, int>(other.F, other.minDst, -other.time, other.i, other.j);
+        return std::tuple<double, int, int, int, int>(F, minDst, -time, i, j) <
+                std::tuple<double, int, int, int, int>(other.F, other.minDst, -other.time, other.i, other.j);
     }
 };
